requirements.txt
fastapi
uvicorn
pymongo
pydantic
transformers
torch
torchvision
pillow
requests
ultralytics
timm
